commit Booking

+++ order/src/main/java/ru/booking/order/api/model/Catalog.java
@@ -1,4 +1,4 @@
 package ru.booking.order.api.model;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 @JsonIgnoreProperties(ignoreUnknown = true)
-public record Catalog(Long id, Long priceForDay, String city, String address) { }
+public record Catalog(Long id, Float priceForDay, String city, String address) { }
